+++ pkg/ccl/importccl/BUILD.bazel
@@ -4,8 +4,10 @@ go_library(
     name = "importccl",
     srcs = [
         "exportcsv.go",
+        "import_job.go",
+        "import_planning.go",
         "import_processor.go",
-        "import_stmt.go",
+        "import_processor_planning.go",
         "import_table_creation.go",
         "import_type_resolver.go",
         "read_import_avro.go",
@@ -60,6 +62,7 @@ go_library(
         "//pkg/sql/pgwire/pgcode",
         "//pkg/sql/pgwire/pgerror",
         "//pkg/sql/pgwire/pgnotice",
+        "//pkg/sql/physicalplan",
         "//pkg/sql/privilege",
         "//pkg/sql/row",
         "//pkg/sql/rowenc",
@@ -88,6 +91,7 @@ go_library(
         "//pkg/workload",
         "@com_github_cockroachdb_apd_v2//:apd",
         "@com_github_cockroachdb_errors//:errors",
+        "@com_github_cockroachdb_logtags//:logtags",
         "@com_github_lib_pq//oid",
         "@com_github_linkedin_goavro_v2//:goavro",
         "@io_vitess_vitess//go/sqltypes",
