+++ pkg/sql/distsql_running.go
@@ -710,6 +710,80 @@ func (w *errOnlyResultWriter) IncrementRowsAffected(ctx context.Context, n int) 
 	panic("IncrementRowsAffected not supported by errOnlyResultWriter")
 }
 
+// RowResultWriter is a thin wrapper around a RowContainer.
+type RowResultWriter struct {
+	rowContainer *rowContainerHelper
+	rowsAffected int
+	err          error
+}
+
+var _ rowResultWriter = &RowResultWriter{}
+
+// NewRowResultWriter creates a new RowResultWriter.
+func NewRowResultWriter(rowContainer *rowContainerHelper) *RowResultWriter {
+	return &RowResultWriter{rowContainer: rowContainer}
+}
+
+// IncrementRowsAffected implements the rowResultWriter interface.
+func (b *RowResultWriter) IncrementRowsAffected(ctx context.Context, n int) {
+	b.rowsAffected += n
+}
+
+// AddRow implements the rowResultWriter interface.
+func (b *RowResultWriter) AddRow(ctx context.Context, row tree.Datums) error {
+	if b.rowContainer != nil {
+		return b.rowContainer.addRow(ctx, row)
+	}
+	return nil
+}
+
+// SetError is part of the rowResultWriter interface.
+func (b *RowResultWriter) SetError(err error) {
+	b.err = err
+}
+
+// Err is part of the rowResultWriter interface.
+func (b *RowResultWriter) Err() error {
+	return b.err
+}
+
+// CallbackResultWriter is a rowResultWriter that runs a callback function
+// on AddRow.
+type CallbackResultWriter struct {
+	fn           func(ctx context.Context, row tree.Datums) error
+	rowsAffected int
+	err          error
+}
+
+var _ rowResultWriter = &CallbackResultWriter{}
+
+// NewCallbackResultWriter creates a new CallbackResultWriter.
+func NewCallbackResultWriter(
+	fn func(ctx context.Context, row tree.Datums) error,
+) *CallbackResultWriter {
+	return &CallbackResultWriter{fn: fn}
+}
+
+// IncrementRowsAffected is part of the rowResultWriter interface.
+func (c *CallbackResultWriter) IncrementRowsAffected(ctx context.Context, n int) {
+	c.rowsAffected += n
+}
+
+// AddRow is part of the rowResultWriter interface.
+func (c *CallbackResultWriter) AddRow(ctx context.Context, row tree.Datums) error {
+	return c.fn(ctx, row)
+}
+
+// SetError is part of the rowResultWriter interface.
+func (c *CallbackResultWriter) SetError(err error) {
+	c.err = err
+}
+
+// Err is part of the rowResultWriter interface.
+func (c *CallbackResultWriter) Err() error {
+	return c.err
+}
+
 var _ execinfra.RowReceiver = &DistSQLReceiver{}
 var _ execinfra.BatchReceiver = &DistSQLReceiver{}
